feat(JS-2194): remove tika dependency
Match CSV files with TextMatcher when extension is ".csv"
