+++ src/main/resources/magic-all.xml
@@ -3517,9 +3517,6 @@
 		<match-regexp offset="4">ftyp(heic|hevs|heis|hevm|hevc|heim|mif1)</match-regexp>
 	</type>
 	<type>
-		   <match-custom>org.jadice.filetype.matchers.CSVMatcher</match-custom>
-	</type>
-	<type>
 		<mime-type>text/plain</mime-type>
 		<extension>txt</extension>
 		<description>ASCII Text Document</description>
